Mappings

BiochemicalAlgorithms._compute_rotation

BiochemicalAlgorithms.match_points

BiochemicalAlgorithms._compute_rotation — Function

_compute_rotation(
    R::StaticArraysCore.SArray{Tuple{3, 3}, T<:Real, 2},
    _::Type{RMSDMinimizerKabsch}
) -> Any

Computes the rotation matrix by solving the eigenvalue problem given as the correlation matrix C. Uses all resulting eigenvalues and eigenvectors. Warns if the correlation matrix is not positive definit (contains negative eigenvalues or eigenvalues equal to 0) and uses the alternative approch RMSDMinimizerCoutsias instead. Returns a RotMatrix3.

_compute_rotation(
    R::StaticArraysCore.SArray{Tuple{3, 3}, T<:Real, 2},
    _::Type{RMSDMinimizerCoutsias}
) -> Rotations.RotMatrix3{T} where T<:Real

Computes the rotation matrix by solving the eigenvalue problem given as the residual matrix F. Uses only the largest of the resulting eigenvalues to generate the quaternion describing the optimal rotation that maps the atoms onto each other. Returns a RotMatrix3.

BiochemicalAlgorithms.match_points — Function

match_points(
    w1::StaticArraysCore.SArray{Tuple{3}, T<:Real, 1, 3},
    w2::StaticArraysCore.SArray{Tuple{3}, T<:Real, 1, 3},
    w3::StaticArraysCore.SArray{Tuple{3}, T<:Real, 1, 3},
    v1::StaticArraysCore.SArray{Tuple{3}, T<:Real, 1, 3},
    v2::StaticArraysCore.SArray{Tuple{3}, T<:Real, 1, 3},
    v3::StaticArraysCore.SArray{Tuple{3}, T<:Real, 1, 3}
) -> Tuple{Any, Any}

The transformation maps

the point w1 onto the point v1 and

the point w2 onto the ray that starts in v1 and goes through v2

the point w3 into the plane generated by v1, v2 and v3