Model

BiochemicalAlgorithms._default_system

BiochemicalAlgorithms._SystemAtomTuple

BiochemicalAlgorithms._SystemChainTuple

BiochemicalAlgorithms._SystemFragmentTuple

BiochemicalAlgorithms._SystemNucleotideTuple

BiochemicalAlgorithms._SystemResidueTuple

BiochemicalAlgorithms._atoms

BiochemicalAlgorithms._bonds

BiochemicalAlgorithms._chains

BiochemicalAlgorithms._fragments

BiochemicalAlgorithms._molecules

BiochemicalAlgorithms._next_idx

BiochemicalAlgorithms._nucleotides

BiochemicalAlgorithms._residues

BiochemicalAlgorithms._row_by_idx

BiochemicalAlgorithms._SystemAtomTuple — Type

const _SystemAtomTuple{T} = NamedTuple{...}

System-specific extension of AtomTuple{T}. See AtomTuple.

Additional fields

frame_id::Int

molecule_id::MaybeInt

chain_id::MaybeInt

fragment_id::MaybeInt

nucleotide_id::MaybeInt

residue_id::MaybeInt

BiochemicalAlgorithms._SystemChainTuple — Type

const _SystemChainTuple{T} = NamedTuple{...}

System-specific extension of ChainTuple{T}. See ChainTuple.

Additional fields

molecule_id::Int

BiochemicalAlgorithms._SystemFragmentTuple — Type

const _SystemFragmentTuple{T} = NamedTuple{...}

System-specific extension of FragmentTuple{T}. See FragmentTuple.

Additional fields

molecule_id::Int

chain_id::Int

BiochemicalAlgorithms._SystemNucleotideTuple — Type

const _SystemNucleotideTuple{T} = NamedTuple{...}

System-specific extension of NucleotideTuple{T}. See NucleotideTuple.

Additional fields

molecule_id::Int

chain_id::Int

BiochemicalAlgorithms._SystemResidueTuple — Type

const _SystemResidueTuple{T} = NamedTuple{...}

System-specific extension of ResidueTuple{T}. See ResidueTuple.

Additional fields

molecule_id::Int

chain_id::Int

BiochemicalAlgorithms._atoms — Function

_atoms(
    sys::System{T};
    frame_id,
    molecule_id,
    chain_id,
    fragment_id,
    nucleotide_id,
    residue_id
) -> DataFrames.SubDataFrame{DataFrames.DataFrame, DataFrames.Index, Vector{Int64}}

Returns a raw DataFrame for all of the given system's atoms matching the given criteria. Fields given as nothing are ignored. Use Some(nothing) if the field should be explicitly checked for a value of nothing. The returned DataFrame contains all public and private atom fields.

_atoms(
    substruct::Substructure{T};
    frame_id,
    molecule_id,
    chain_id,
    fragment_id,
    nucleotide_id,
    residue_id
) -> DataFrames.SubDataFrame{DataFrames.DataFrame, DataFrames.Index, Vector{Int64}}

Returns a raw DataFrame for all of the given system's atoms matching the given criteria (value or missing). Fields given as nothing are ignored. The returned DataFrame contains all public and private atom fields.

BiochemicalAlgorithms._bonds — Function

_bonds(
    sys::System;
    kwargs...
) -> DataFrames.SubDataFrame{DataFrames.DataFrame, DataFrames.Index}

Returns a raw DataFrame for all of the given system's bonds associated with at least one atom matching the given criteria. Fields given as nothing are ignored. Use Some(nothing) if the field should be explicitly checked for a value of nothing. See _atoms.

_bonds(
    atom::Atom
) -> DataFrames.SubDataFrame{DataFrames.DataFrame, DataFrames.Index}

Returns a raw DataFrame containing all of the given atom's bonds.

BiochemicalAlgorithms._chains — Function

_chains(sys::System; molecule_id) -> DataFrames.DataFrame

Returns a raw DataFrame for all of the given system's chains matching the given criteria. Fields given as nothing are ignored. The returned DataFrame contains all public and private chain fields.

BiochemicalAlgorithms._default_system — Constant

const _default_system

Global default system.

BiochemicalAlgorithms._fragments — Function

_fragments(
    sys::System{T};
    molecule_id,
    chain_id
) -> DataFrames.DataFrame

Returns a raw DataFrame for all of the given system's fragments matching the given criteria. Fields given as nothing are ignored. The returned DataFrame contains all public and private fragment fields.

BiochemicalAlgorithms._molecules — Function

_molecules(sys::System) -> DataFrames.DataFrame

Returns a raw DataFrame for all of the given system's molecules. The returned DataFrame contains all public and private molecule fields.

BiochemicalAlgorithms._nucleotides — Function

_nucleotides(
    sys::System{T};
    molecule_id,
    chain_id
) -> DataFrames.DataFrame

Returns a raw DataFrame for all of the given system's nucleotides matching the given criteria. Fields given as nothing are ignored. The returned DataFrame contains all public and private nucleotide fields.

BiochemicalAlgorithms._next_idx — Function

_next_idx(sys::System{T}) -> Int64

Returns the next available idx for the given system.

BiochemicalAlgorithms._residues — Function

_residues(
    sys::System{T};
    molecule_id,
    chain_id
) -> DataFrames.DataFrame

Returns a raw DataFrame for all of the given system's residues matching the given criteria. Fields given as nothing are ignored. The returned DataFrame contains all public and private residue fields.

BiochemicalAlgorithms._row_by_idx — Function

_row_by_idx(df::DataFrames.DataFrame, idx::Int64) -> Int64

Returns the row number corresponding to the given idx in df. Throws a KeyError if no such row exists.