File formats

BiochemicalAlgorithms.load_hinfile

BiochemicalAlgorithms.load_mmcif

BiochemicalAlgorithms.load_pdb

BiochemicalAlgorithms.load_pubchem_json

BiochemicalAlgorithms.load_sdfile

BiochemicalAlgorithms.read_ball_ini_file

BiochemicalAlgorithms.write_hinfile

BiochemicalAlgorithms.write_mmcif

BiochemicalAlgorithms.write_pdb

BiochemicalAlgorithms.write_sdfile

BALL INI files

BiochemicalAlgorithms.read_ball_ini_file — Function

read_ball_ini_file(fname::AbstractString, ::Type{T} = Float32) -> BALLIniFile

Read a file in BALL's old Ini file format and return it as a BALLIniFile.

Supported keyword arguments

cleanup_keys::Bool = true simplifies colon-separated key names (e.g., ver:version becomes version, key:I becomes I, etc.)

HyperChem HIN files

BiochemicalAlgorithms.load_hinfile — Function

load_hinfile(
    fname::AbstractString,
    ::Type{T} = Float32
) -> System{T}

Read a HyperChem HIN file.

BiochemicalAlgorithms.write_hinfile — Function

write_hinfile(
    fname::AbstractString,
    ac::AtomContainer
)

Save an atom container as HyperChem HIN file.

Note

HIN files define molecules as connected components in the molecular graph. If the AtomContainer is missing bonds, e.g., after reading a PDB file and not postprocessing it correctly, the HIN file may contain a surprisingly large number of molecules.

PDB files

BiochemicalAlgorithms.load_pdb — Function

load_pdb(fname::AbstractString, ::Type{T} = Float32) -> System{T}

Read a PDB file.

Note

Models are stored as frames, using the model number as frame_id.

BiochemicalAlgorithms.write_pdb — Function

write_pdb(fname::AbstractString, ac::AbstractAtomContainer)

Save an atom container as PDB file.

PDBx/mmCIF files

BiochemicalAlgorithms.load_mmcif — Function

load_mmcif(fname::AbstractString, ::Type{T} = Float32) -> System{T}

Read a PDBx/mmCIF file.

Note

Models are stored as frames, using the model number as frame_id.

BiochemicalAlgorithms.write_mmcif — Function

write_mmcif(fname::AbstractString, ac::AbstractAtomContainer)

Save an atom container as PDBx/mmCIF file.

PubChem JSON

BiochemicalAlgorithms.load_pubchem_json — Function

load_pubchem_json(fname::AbstractString, ::Type{T} = Float32) -> System{T}

Read a PubChem JSON file.

Note

Conformers are stored as frames.

Structures data files (SDF, SDFile)

BiochemicalAlgorithms.load_sdfile — Function

load_sdfile(fname::AbstractString, ::Type{T} = Float32) -> System{T}

Read an SDFile.

BiochemicalAlgorithms.write_sdfile — Function

write_sdfile(fname::AbstractString, ac::AbstractAtomContainer)

Save a 2D projection of an atom container as SDFile.