Force fields

BiochemicalAlgorithms.AmberFF

BiochemicalAlgorithms.optimize_hydrogen_positions!

BiochemicalAlgorithms.optimize_structure!

BiochemicalAlgorithms.update!

BiochemicalAlgorithms.AmberFF — Function

AmberFF(
    ::AbstractAtomContainer{T},
    param_file::AbstractString = ball_data_path("forcefields/AMBER/amber96.ini")
)

Initializes an AMBER force field for the given atom container and the given parameter file (default: AMBER96).

Supported keyword arguments

nonbonded_cutoff::T = 20

vdw_cutoff::T = 15

vdw_cuton::T = 13

electrostatic_cutoff::T = 15

electrostatic_cuton::T = 13

scaling_vdw_1_4 = 2

scaling_electrostatic_1_4::T = 1.2

distance_dependent_dielectric::Bool = false

assign_charges::Bool = true

assign_typenames::Bool = true

assign_types::Bool = true

overwrite_nonzero_charges::Bool = true

overwrite_typenames::Bool = false

periodic_boundary_conditions::Bool = false

periodic_box_width::T = 100

periodic_box_height::T = 100

periodic_box_depth::T = 100

max_number_of_unassigned_atoms::Int = typemax(Int32)

BiochemicalAlgorithms.optimize_hydrogen_positions! — Function

optimize_hydrogen_positions!(ff::ForceField)

Variant of optimize_structure! that only optimizes hydrogen atom positions.

Supported keyword arguments

Same as optimize_structure!

BiochemicalAlgorithms.optimize_structure! — Function

optimize_structure!(ff::ForceField)

Attempts to solve the energy optimization problem represented by the given force field object.

Supported keyword arguments

This function passes all keyword arguments to Optimization.solve, with the following default values:

alg = Optimization.LBFGS()

BiochemicalAlgorithms.update! — Function

update!(ff::ForceField)

Update the internal data structures of the force field when the system changes (e.g., through coordinate updates).

Note

Changes to the options or the topology require a call to setup! prior to the call to update!.