Force fields

BiochemicalAlgorithms.optimize_hydrogen_positions!

BiochemicalAlgorithms.optimize_structure!

BiochemicalAlgorithms.read_ball_ini_file

BiochemicalAlgorithms.update!

BiochemicalAlgorithms.optimize_hydrogen_positions! — Function

optimize_hydrogen_positions!(ff::ForceField)

Variant of optimize_structure! that only optimizes hydrogen atom positions.

Supported keyword arguments

Same as optimize_structure!

BiochemicalAlgorithms.optimize_structure! — Function

optimize_structure!(ff::ForceField)

Attempts to solve the energy optimization problem represented by the given force field object.

Supported keyword arguments

This function passes all keyword arguments to Optimization.solve, with the following default values:

alg = Optimization.LBFGS()

BiochemicalAlgorithms.read_ball_ini_file — Function

read_ball_ini_file(path::String; ...) -> BALLIniFile
read_ball_ini_file(
    path::String,
    T;
    cleanup_keys
) -> BALLIniFile

Read a file in BALL's old Ini file format and return it as a BALLIniFile.

If cleanup_keys is set to true (the default), keys into the Ini sections are simplified (e.g., ver:version becomes version, key:I becomes I, ...).

BiochemicalAlgorithms.update! — Function

update!(ff::ForceField)

Update the internal data structures of the force field when the system changes (e.g., through coordinate updates).

Note

Changes to the options or the topology require a call to setup! prior to the call to update!.