System representation

BiochemicalAlgorithms.AbstractAtomContainer

BiochemicalAlgorithms.AbstractMolecule

BiochemicalAlgorithms.AbstractSystemComponent

BiochemicalAlgorithms.Atom

BiochemicalAlgorithms.AtomTuple

BiochemicalAlgorithms.Bond

BiochemicalAlgorithms.BondTuple

BiochemicalAlgorithms.Chain

BiochemicalAlgorithms.ChainTuple

BiochemicalAlgorithms.Fragment

BiochemicalAlgorithms.FragmentTuple

BiochemicalAlgorithms.Molecule

BiochemicalAlgorithms.MoleculeTuple

BiochemicalAlgorithms.Nucleotide

BiochemicalAlgorithms.NucleotideTuple

BiochemicalAlgorithms.Residue

BiochemicalAlgorithms.ResidueTuple

BiochemicalAlgorithms.System

Base.parent

Base.push!

Base.push!

Base.push!

Base.push!

Base.push!

BiochemicalAlgorithms.atom_by_idx

BiochemicalAlgorithms.atoms

BiochemicalAlgorithms.atoms_df

BiochemicalAlgorithms.bond_by_idx

BiochemicalAlgorithms.bonds

BiochemicalAlgorithms.bonds_df

BiochemicalAlgorithms.chain_by_idx

BiochemicalAlgorithms.chains

BiochemicalAlgorithms.chains_df

BiochemicalAlgorithms.default_system

BiochemicalAlgorithms.eachatom

BiochemicalAlgorithms.eachbond

BiochemicalAlgorithms.eachchain

BiochemicalAlgorithms.eachfragment

BiochemicalAlgorithms.eachmolecule

BiochemicalAlgorithms.eachnucleotide

BiochemicalAlgorithms.eachresidue

BiochemicalAlgorithms.fragment_by_idx

BiochemicalAlgorithms.fragments

BiochemicalAlgorithms.fragments_df

BiochemicalAlgorithms.get_property

BiochemicalAlgorithms.has_flag

BiochemicalAlgorithms.has_property

BiochemicalAlgorithms.is_bound_to

BiochemicalAlgorithms.is_geminal

BiochemicalAlgorithms.is_vicinal

BiochemicalAlgorithms.molecule_by_idx

BiochemicalAlgorithms.molecules

BiochemicalAlgorithms.molecules_df

BiochemicalAlgorithms.natoms

BiochemicalAlgorithms.nbonds

BiochemicalAlgorithms.nchains

BiochemicalAlgorithms.nfragments

BiochemicalAlgorithms.nmolecules

BiochemicalAlgorithms.nnucleotides

BiochemicalAlgorithms.nresidues

BiochemicalAlgorithms.nucleotide_by_idx

BiochemicalAlgorithms.nucleotides

BiochemicalAlgorithms.nucleotides_df

BiochemicalAlgorithms.parent_chain

BiochemicalAlgorithms.parent_fragment

BiochemicalAlgorithms.parent_molecule

BiochemicalAlgorithms.parent_nucleotide

BiochemicalAlgorithms.parent_residue

BiochemicalAlgorithms.parent_system

BiochemicalAlgorithms.residue_by_idx

BiochemicalAlgorithms.residues

BiochemicalAlgorithms.residues_df

BiochemicalAlgorithms.set_flag!

BiochemicalAlgorithms.set_property!

BiochemicalAlgorithms.unset_flag!

Abstract types

BiochemicalAlgorithms.AbstractSystemComponent — Type

abstract type AbstractSystemComponent{T<:Real}

Abstract base type for all components of a system, including the system itself.

BiochemicalAlgorithms.AbstractAtomContainer — Type

abstract type AbstractAtomContainer{T} <: AbstractSystemComponent{T}

Abstract base type for all atom containers.

BiochemicalAlgorithms.AbstractMolecule — Type

abstract type AbstractMolecule{T} <: AbstractAtomContainer{T}

Abstract base type for all molecules.

Common functions

BiochemicalAlgorithms.has_property — Function

has_property(
    ac::AbstractSystemComponent,
    key::Symbol
) -> Any

Returns a Bool indicating whether the given system component has the given property.

BiochemicalAlgorithms.get_property — Function

get_property(
    ac::AbstractSystemComponent,
    key::Symbol
) -> Any

Returns the property associated with the given key in ac.

get_property(
    ac::AbstractSystemComponent,
    key::Symbol,
    default
) -> Any

Returns the property associated with the given key in ac. If no such property exists, returns default.

BiochemicalAlgorithms.set_property! — Function

set_property!(
    ac::AbstractSystemComponent,
    key::Symbol,
    value
) -> Any

Sets the property associated with the given key in ac to the given value.

BiochemicalAlgorithms.has_flag — Function

has_flag(ac::AbstractSystemComponent, flag::Symbol) -> Any

Returns a Bool indicating whether the given system component has the given flag.

BiochemicalAlgorithms.set_flag! — Function

set_flag!(ac::AbstractSystemComponent, flag::Symbol) -> Any

Adds the given flag to ac.

BiochemicalAlgorithms.unset_flag! — Function

unset_flag!(
    ac::AbstractSystemComponent,
    flag::Symbol
) -> Any

Removes the given flag from ac.

Systems

BiochemicalAlgorithms.System — Type

mutable struct System{T} <: AbstractAtomContainer{T}

Mutable representation of a biomolecular system.

Fields

name::String

properties::Properties

flags::Flags

Constructors

System(name::String = "", properties::Properties = Properties(), flags::Flags = Flags())

Creates a new and empty System{Float32}.

System{T}(name::String = "", properties::Properties = Properties(), flags::Flags = Flags())

Creates a new and empty System{T}.

BiochemicalAlgorithms.default_system — Function

default_system() -> System{Float32}

Returns the global default system.

Base.parent — Method

parent(::Atom)
parent(::Bond)
parent(::Chain)
parent(::Fragment)
parent(::Molecule)
parent(::Nucleotide)
parent(::Protein)
parent(::Residue)
parent(::System)

Returns the System{T} containing the given object.

BiochemicalAlgorithms.parent_system — Function

parent_system(::Atom)
parent_system(::Bond)
parent_system(::Chain)
parent_system(::Fragment)
parent_system(::Molecule)
parent_system(::Nucleotide)
parent_system(::Protein)
parent_system(::Residue)
parent_system(::System)

Returns the System{T} containing the given object. Alias for Base.parent.

Atoms

BiochemicalAlgorithms.Atom — Type

struct Atom{T} <: AbstractSystemComponent{T}

Mutable representation of an individual atom in a system.

Fields

idx::Int

number::Int

element::ElementType

name::String

atom_type::String

r::Vector3{T}

v::Vector3{T}

F::Vector3{T}

formal_charge::Int

charge::T

radius::T

properties::Properties

flags::Flags

Constructors

Atom(
    number::Int,
    element::ElementType,
    name::String = "",
    atom_type::String = "",
    r::Vector3{T} = Vector3{T}(0, 0, 0),
    v::Vector3{T} = Vector3{T}(0, 0, 0),
    F::Vector3{T} = Vector3{T}(0, 0, 0),
    formal_charge::Int = 0,
    charge::T = zero(T),
    radius::T = zero(T),
    properties::Properties = Properties(),
    flags::Flags = Flags();
    # keyword arguments
    frame_id::Int = 1
)

Creates a new Atom{Float32} in the default system.

Atom(
    ac::AbstractAtomContainer{T},
    number::Int,
    element::ElementType,
    name::String = "",
    atom_type::String = "",
    r::Vector3{T} = Vector3{T}(0, 0, 0),
    v::Vector3{T} = Vector3{T}(0, 0, 0),
    F::Vector3{T} = Vector3{T}(0, 0, 0),
    formal_charge::Int = 0,
    charge::T = zero(T),
    radius::T = zero(T),
    properties::Properties = Properties(),
    flags::Flags = Flags();
    # keyword arguments
    frame_id::Int = 1
)

Creates a new Atom{T} in the given atom container (e.g., System{T} or Molecule{T}).

Atom(a::AtomTuple{T}; frame_id::Int = 1)
Atom(sys::System{T}, a::AtomTuple{T}; frame_id::Int = 1)

Constructor variants that create a new system atom based on the given AtomTuple{T}. The new atom is automatically assigned a new idx.

BiochemicalAlgorithms.AtomTuple — Type

const AtomTuple{T} = NamedTuple{...}

Tuple-based atom representation for DataFrame usage.

Fields

idx::Int

number::Int

element::ElementType

name::String

atom_type::String

r::Vector3{T}

v::Vector3{T}

F::Vector3{T}

properties::Properties

flags::Flags

Constructors

AtomTuple(
    number::Int,
    element::ElementType;
    # keyword arguments
    idx::Int = 0,
    name::String = "",
    atom_type::String = "",
    r::Vector3{T} = Vector3{T}(0, 0, 0),
    v::Vector3{T} = Vector3{T}(0, 0, 0),
    F::Vector3{T} = Vector3{T}(0, 0, 0),
    formal_charge::Int = 0,
    charge::T = zero(T),
    radius::T = zero(T),
    properties::Properties = Properties(),
    flags::Flags = Flags()
)

Creates a new AtomTuple{Float32} with default values for all omitted fields.

AtomTuple{T}(
    number::Int,
    element::ElementType;
    # keyword arguments
    idx::Int = 0,
    name::String = "",
    atom_type::String = "",
    r::Vector3{T} = Vector3{T}(0, 0, 0),
    v::Vector3{T} = Vector3{T}(0, 0, 0),
    F::Vector3{T} = Vector3{T}(0, 0, 0),
    formal_charge::Int = 0,
    charge::T = zero(T),
    radius::T = zero(T),
    properties::Properties = Properties(),
    flags::Flags = Flags()
)

Creates a new AtomTuple{T} with default values for all omitted fields.

BiochemicalAlgorithms.atom_by_idx — Function

atom_by_idx(sys::System{T}, idx::Int64) -> Atom

Returns the Atom{T} associated with the given idx in sys. Returns nothing if no such atom exists.

BiochemicalAlgorithms.atoms — Function

atoms(::Chain)
atoms(::Fragment)
atoms(::Molecule)
atoms(::Nucleotide)
atoms(::Protein)
atoms(::Residue)
atoms(::System)

Returns a Vector{Atom{T}} containing all atoms of the given atom container.

Supported keyword arguments

frame_id::MaybeInt = 1: Any value other than nothing limits the result to atoms matching this frame ID.

BiochemicalAlgorithms.atoms_df — Function

atoms_df(::Chain)
atoms_df(::Fragment)
atoms_df(::Molecule)
atoms_df(::Nucleotide)
atoms_df(::Protein)
atoms_df(::Residue)
atoms_df(::System)

Returns a SubDataFrame containing all atoms of the given atom container.

Supported keyword arguments

frame_id::MaybeInt = 1: Any value other than nothing limits the result to atoms matching this frame ID.

BiochemicalAlgorithms.eachatom — Function

eachatom(::Chain)
eachatom(::Fragment)
eachatom(::Molecule)
eachatom(::Nucleotide)
eachatom(::Protein)
eachatom(::Residue)
eachatom(::System)

Returns an Atom{T} generator for all atoms of the given atom container.

Supported keyword arguments

frame_id::MaybeInt = 1: Any value other than nothing limits the result to atoms matching this frame ID.

Example

for atom in eachatom(sys)
    println(atom.name)
end

BiochemicalAlgorithms.is_bound_to — Function

is_bound_to(a1::Atom, a2::Atom) -> Bool

Decides if two atoms are bound to each other.
Hydrogen bonds (has_flag(bond, :TYPE__HYDROGEN)) are ignored.

BiochemicalAlgorithms.is_geminal — Function

is_geminal(a1::Atom, a2::Atom) -> Any

Decides if two atoms are geminal.

Two atoms are geminal if they do not share a common bond but both have a
bond to a third atom. For example the two hydrogen atoms in water are geminal. 
Hydrogen bonds (has_flag(bond, :TYPE__HYDROGEN)) are ignored.

BiochemicalAlgorithms.is_vicinal — Function

is_vicinal(a1::Atom, a2::Atom) -> Bool

Decides if two atoms are vicinal.

Two atoms are vicinal if they are separated by three bonds (1-4 position). Hydrogen bonds (has_flag(bond, :TYPE__HYDROGEN)) are ignored.

BiochemicalAlgorithms.natoms — Function

natoms(::Chain)
natoms(::Fragment)
natoms(::Molecule)
natoms(::Nucleotide)
natoms(::Protein)
natoms(::Residue)
natoms(::System)

Returns the number of atoms in the given atom container.

Supported keyword arguments

frame_id::MaybeInt = 1: Any value other than nothing limits the result to atoms matching this frame ID.

Base.push! — Method

push!(::Fragment{T}, atom::AtomTuple{T})
push!(::Molecule{T}, atom::AtomTuple{T})
push!(::Nucleotide{T}, atom::AtomTuple{T})
push!(::Protein{T}, atom::AtomTuple{T})
push!(::Residue{T}, atom::AtomTuple{T})
push!(::System{T}, atom::AtomTuple{T})

Creates a new atom in the given atom container, based on the given tuple. The new atom is automatically assigned a new idx.

Supported keyword arguments

frame_id::Int = 1

Bonds

BiochemicalAlgorithms.Bond — Type

struct Bond{T} <: AbstractSystemComponent{T}

Mutable representation of an individual bond in a system.

Fields

idx::Int

a1::Int

a2::Int

order::BondOrderType

properties::Properties

flags::Flags

Constructors

Bond(
    a1::Int, 
    a2::Int, 
    order::BondOrderType, 
    properties::Properties = Properties(),
    flags::Flags = Flags()
)

Creates a new Bond{Float32} in the default system.

Bond(
    sys::System{T}, 
    a1::Int, 
    a2::Int, 
    order::BondOrderType, 
    properties::Properties = Properties(),
    flags::Flags = Flags()
)

Creates a new Bond{T} in the given system.

BiochemicalAlgorithms.BondTuple — Type

const BondTuple = NamedTuple{...}

Tuple-based bond representation for DataFrame usage.

Fields

idx::Int

a1::Int

a2::Int

order::BondOrderType

properties::Properties

flags::Flags

Constructors

BondTuple(
    a1::Int,
    a2::Int,
    order::BondOrderType;
    # keyword arguments
    idx::Int = 0,
    properties::Properties = Properties(),
    flags::Flags = Flags()
)

Creates a new BondTuple with default values for all omitted fields.

BiochemicalAlgorithms.bond_by_idx — Function

bond_by_idx(
    sys::System{T},
    idx::Int64
) -> Union{Nothing, Bond}

Returns the Bond{T} associated with the given idx in sys. Returns nothing if no such bond exists.

BiochemicalAlgorithms.bonds — Function

bonds(::Chain)
bonds(::Fragment)
bonds(::Molecule)
bonds(::Nucleotide)
bonds(::Protein)
bonds(::Residue)
bonds(::System)

Returns a Vector{Bond{T}} containing all bonds of the given atom container where at least one associated atom is contained in the same container.

Supported keyword arguments

frame_id::MaybeInt = 1: Any value other than nothing also limits the result to bonds where at least on atom matches this frame ID.

bonds(::Atom)

Returns a Vector{Bond{T}} containing all bonds of the given atom.

BiochemicalAlgorithms.bonds_df — Function

bonds_df(::Chain)
bonds_df(::Fragment)
bonds_df(::Molecule)
bonds_df(::Nucleotide)
bonds_df(::Protein)
bonds_df(::Residue)
bonds_df(::System)

Returns a SubDataFrame containing all bonds of the given atom container where at least one associated atom is contained in the same container.

Supported keyword arguments

frame_id::MaybeInt = 1: Any value other than nothing also limits the result to bonds where at least on atom matches this frame ID.

bonds_df(::Atom)

Returns a SubDataFrame containing all bonds of the given atom.

BiochemicalAlgorithms.eachbond — Function

eachbond(::Chain)
eachbond(::Fragment)
eachbond(::Molecule)
eachbond(::Nucleotide)
eachbond(::Protein)
eachbond(::Residue)
eachbond(::System)

Returns a Bond{T} generator for all bonds of the given atom container where at least one associated atom is contained in the same container.

Supported keyword arguments

frame_id::MaybeInt = 1: Any value other than nothing also limits the result to bonds where at least on atom matches this frame ID.

eachbond(::Atom)

Returns a Bond{T} generator for all bonds of the given atom.

BiochemicalAlgorithms.nbonds — Function

nbonds(::Chain)
nbonds(::Fragment)
nbonds(::Molecule)
nbonds(::Nucleotide)
nbonds(::Protein)
nbonds(::Residue)
nbonds(::System)

Returns the number of bonds in the given atom container where at least one associated atom is contained in the same container.

Supported keyword arguments

frame_id::MaybeInt = 1: Any value other than nothing also limits the result to bonds where at least on atom matches this frame ID.

nbonds(::Atom)

Returns the number of bonds of the given atom.

Base.push! — Method

push!(ac::AbstractAtomContainer, bond::BondTuple)

Creates a new bond in the system associated with the given atom container, based on the given tuple. The new bond is automatically assigned a new idx.

Molecules

BiochemicalAlgorithms.Molecule — Type

struct Molecule{T} <: AbstractMolecule{T}

Mutable representation of an individual molecule in a system.

Fields

idx::Int

name::String

properties::Properties

flags::Flags

Constructors

Molecule(
    name::String = "",
    properties::Properties = Properties(),
    flags::Flags = Flags()
)

Creates a new Molecule{Float32} in the default system.

Molecule(
    sys::System{T},
    name::String = "",
    properties::Properties = Properties(),
    flags::Flags = Flags()
)

Creates a new Molecule{T} in the given system.

BiochemicalAlgorithms.MoleculeTuple — Type

const MoleculeTuple = NamedTuple{...}

Tuple-based molecule representation for DataFrame usage.

Fields

idx::Int

name::String

properties::Properties

flags::Flags

Constructors

MoleculeTuple(;
    # keyword arguments
    idx::Int = 0,
    name::String = "",
    properties::Properties = Properties(),
    flags::Flags = Flags()
)

Creates a new MoleculeTuple with default values for all omitted fields.

BiochemicalAlgorithms.eachmolecule — Function

eachmolecule(
    sys::System{T}
) -> Base.Generator{DataFrames.DataFrameRows{DataFrames.DataFrame}}

Returns a Molecule{T} generator for all molecules of the given system.

BiochemicalAlgorithms.molecule_by_idx — Function

molecule_by_idx(
    sys::System{T},
    idx::Int64
) -> Union{Nothing, Molecule}

Returns the Molecule{T} associated with the given idx in sys. Returns nothing if no such molecule exists.

BiochemicalAlgorithms.molecules — Function

molecules(sys::System) -> Vector

Returns a Vector{Molecule{T}} containing all molecules of the given system.

BiochemicalAlgorithms.molecules_df — Function

molecules_df(
    sys::System{T}
) -> DataFrames.SubDataFrame{DataFrames.DataFrame, DataFrames.Index, Base.OneTo{Int64}}

Returns a SubDataFrame containing all molecules of the given system.

BiochemicalAlgorithms.nmolecules — Function

nmolecules(sys::System) -> Int64

Returns the number of molecules in the given system.

BiochemicalAlgorithms.parent_molecule — Function

parent_molecule(::Atom)
parent_molecule(::Chain)
parent_molecule(::Fragment)
parent_molecule(::Nucleotide)
parent_molecule(::Residue)

Returns the Molecule{T} containing the given object. Returns nothing if no such molecule exists.

Chains

BiochemicalAlgorithms.Chain — Type

struct Chain{T} <: AbstractAtomContainer{T}

Mutable representation of an individual chain in a system.

Fields

idx::Int

name::String

properties::Properties

flags::Flags

Constructors

Chain(
    mol::Molecule{T},
    name::String = "",
    properties::Properties = Properties(),
    flags::Flags = Flags()
)

Creates a new Chain{T} in the given molecule.

BiochemicalAlgorithms.ChainTuple — Type

const ChainTuple = NamedTuple{...}

Tuple-based chain representation for DataFrame usage.

Fields

idx::Int

name::String

properties::Properties

flags::Flags

Constructors

ChainTuple(;
    # keyword arguments
    idx::Int = 0,
    name::String = "",
    properties::Properties = Properties(),
    flags::Flags = Flags()
)

Creates a new ChainTuple with default values for all omitted fields.

BiochemicalAlgorithms.chain_by_idx — Function

chain_by_idx(
    sys::System{T},
    idx::Int64
) -> Union{Nothing, Chain}

Returns the Chain{T} associated with the given idx in sys. Returns nothing if no such chain exists.

BiochemicalAlgorithms.chains — Function

chains(::Molecule)
chains(::Protein)
chains(::System)

Returns a Vector{Chain{T}} containing all chains of the given atom container.

BiochemicalAlgorithms.chains_df — Function

chains_df(::Molecule)
chains_df(::Protein)
chains_df(::System)

Returns a SubDataFrame containing all chains of the given atom container.

BiochemicalAlgorithms.eachchain — Function

eachchain(::Molecule)
eachchain(::Protein)
eachchain(::System)

Returns a Chain{T} generator for all chains of the given atom container.

BiochemicalAlgorithms.nchains — Function

nchains(::Molecule)
nchains(::Protein)
nchains(::System)

Returns the number of chains in the given atom container.

BiochemicalAlgorithms.parent_chain — Function

parent_chain(::Atom)
parent_chain(::Fragment)
parent_chain(::Nucleotide)
parent_chain(::Residue)

Returns the Chain{T} containing the given object. Returns nothing if no such chain exists.

Base.push! — Method

push!(::Molecule, chain::ChainTuple)
push!(::Protein, chain::ChainTuple)

Creates a new chain in the given molecule, based on the given tuple. The new chain is automatically assigned a new idx.

Fragments

BiochemicalAlgorithms.Fragment — Type

struct Fragment{T} <: AbstractAtomContainer{T}

Mutable representation of an individual fragment in a system.

Fields

idx::Int

number::Int

name::String

properties::Properties

flags::Flags

Constructors

Fragment(
    chain::Chain{T},
    number::Int,
    name::String = "",
    properties::Properties = Properties(),
    flags::Flags = Flags()
)

Creates a new Fragment{T} in the given chain.

BiochemicalAlgorithms.FragmentTuple — Type

const FragmentTuple = NamedTuple{...}

Tuple-based fragment representation for DataFrame usage.

Fields

idx::Int

number::Int

name::String

properties::Properties

flags::Flags

Constructors

FragmentTuple(
    number::Int;
    # keyword arguments
    idx::Int = 0,
    name::String = "",
    properties::Properties = Properties(),
    flags::Flags = Flags()
)

Creates a new FragmentTuple with default values for all omitted fields.

BiochemicalAlgorithms.eachfragment — Function

eachfragment(::Chain)
eachfragment(::Molecule)
eachfragment(::Protein)
eachfragment(::System)

Returns a Fragment{T} generator for all fragments of the given atom container.

BiochemicalAlgorithms.fragment_by_idx — Function

fragment_by_idx(
    sys::System{T},
    idx::Int64
) -> Union{Nothing, Fragment}

Returns the Fragment{T} associated with the given idx in sys. Returns nothing if no such fragment exists.

BiochemicalAlgorithms.fragments — Function

fragments(::Chain)
fragments(::Molecule)
fragments(::Protein)
fragments(::System)

Returns a Vector{Fragment{T}} containing all fragments of the given atom container.

BiochemicalAlgorithms.fragments_df — Function

fragments_df(::Chain)
fragments_df(::Molecule)
fragments_df(::Protein)
fragments_df(::System)

Returns a SubDataFrame containing all fragments of the given atom container.

BiochemicalAlgorithms.nfragments — Function

nfragments(::Chain)
nfragments(::Molecule)
nfragments(::Protein)
nfragments(::System)

Returns the number of fragments in the given atom container.

BiochemicalAlgorithms.parent_fragment — Function

parent_fragment(::Atom)

Returns the Fragment{T} containing the given atom. Returns nothing if no such fragment exists.

Base.push! — Method

push!(::Chain, frag::FragmentTuple)

Creates a new fragment in the given chain, based on the given tuple. The new fragment is automatically assigned a new idx.

Nucleotides

BiochemicalAlgorithms.Nucleotide — Type

struct Nucleotide{T} <: AbstractAtomContainer{T}

Mutable representation of an individual nucleotide in a system.

Fields

idx::Int

number::Int

name::String

properties::Properties

flags::Flags

Constructors

Nucleotide(
    chain::Chain{T},
    number::Int,
    name::String = "",
    properties::Properties = Properties(),
    flags::Flags = Flags()
)

Creates a new Nucleotide{T} in the given chain.

BiochemicalAlgorithms.NucleotideTuple — Type

const NucleotideTuple = NamedTuple{...}

Tuple-based nucleotide representation for DataFrame usage.

Fields

idx::Int

number::Int

name::String

properties::Properties

flags::Flags

Constructors

NucleotideTuple(
    number::Int;
    # keyword arguments
    idx::Int = 0,
    name::String = "",
    properties::Properties = Properties(),
    flags::Flags = Flags()
)

Creates a new NucleotideTuple with default values for all omitted fields.

BiochemicalAlgorithms.eachnucleotide — Function

eachnucleotide(::Chain)
eachnucleotide(::Molecule)
eachnucleotide(::Protein)
eachnucleotide(::System)

Returns a Nucleotide{T} generator for all nucleotides of the given atom container.

BiochemicalAlgorithms.nnucleotides — Function

nnucleotides(::Chain)
nnucleotides(::Molecule)
nnucleotides(::Protein)
nnucleotides(::System)

Returns the number of nucleotides in the given atom container.

BiochemicalAlgorithms.nucleotide_by_idx — Function

nucleotide_by_idx(
    sys::System{T},
    idx::Int64
) -> Union{Nothing, Nucleotide}

Returns the Nucleotide{T} associated with the given idx in sys. Returns nothing if no such nucleotide exists.

BiochemicalAlgorithms.nucleotides — Function

nucleotides(::Chain)
nucleotides(::Molecule)
nucleotides(::Protein)
nucleotides(::System)

Returns a Vector{Nucleotide{T}} containing all nucleotides of the given atom container.

BiochemicalAlgorithms.nucleotides_df — Function

nucleotides_df(::Chain)
nucleotides_df(::Molecule)
nucleotides_df(::Protein)
nucleotides_df(::System)

Returns a SubDataFrame containing all nucleotides of the given atom container.

BiochemicalAlgorithms.parent_nucleotide — Function

parent_nucleotide(::Atom)

Returns the Nucleotide{T} containing the given atom. Returns nothing if no such nucleotide exists.

Residues

BiochemicalAlgorithms.Residue — Type

struct Residue{T} <: AbstractAtomContainer{T}

Mutable representation of an individual residue in a system.

Fields

idx::Int

number::Int

type::AminoAcid

properties::Properties

flags::Flags

Constructors

Residue(
    chain::Chain{T},
    number::Int,
    type::AminoAcid,
    properties::Properties = Properties(),
    flags::Flags = Flags()
)

Creates a new Residue{T} in the given chain.

BiochemicalAlgorithms.ResidueTuple — Type

const ResidueTuple = NamedTuple{...}

Tuple-based residue representation for DataFrame usage.

Fields

idx::Int

number::Int

type::AminoAcid

properties::Properties

flags::Flags

Constructors

ResidueTuple(
    number::Int,
    type::AminoAcid;
    # keyword arguments
    idx::Int = 0,
    properties::Properties = Properties(),
    flags::Flags = Flags()
)

BiochemicalAlgorithms.eachresidue — Function

eachresidue(::Chain)
eachresidue(::Molecule)
eachresidue(::Protein)
eachresidue(::System)

Returns a Residue{T} generator for all residues of the given atom container.

BiochemicalAlgorithms.nresidues — Function

nresidues(::Chain)
nresidues(::Molecule)
nresidues(::Protein)
nresidues(::System)

Returns the number of residues in the given atom container.

BiochemicalAlgorithms.parent_residue — Function

parent_residue(::Atom)

Returns the Residue{T} containing the given atom. Returns nothing if no such residue exists.

BiochemicalAlgorithms.residue_by_idx — Function

residue_by_idx(
    sys::System{T},
    idx::Int64
) -> Union{Nothing, Residue}

Returns the Residue{T} associated with the given idx in sys. Returns nothing if no such residue exists.

BiochemicalAlgorithms.residues — Function

residues(::Chain)
residues(::Molecule)
residues(::Protein)
residues(::System)

Returns a Vector{Residue{T}} containing all residues of the given atom container.

BiochemicalAlgorithms.residues_df — Function

residues_df(::Chain)
residues_df(::Molecule)
residues_df(::Protein)
residues_df(::System)

Returns a SubDataFrame{T} containing all residues of the given atom container.

Base.push! — Method

push!(::Chain, frag::FragmentTuple)

Creates a new fragment in the given chain, based on the given tuple. The new residue is automatically assigned a new idx.